requirements.txt
certifi==2021.5.30
chardet==4.0.0
Django==2.2.24
djangorestframework==3.12.4
idna==2.10
pkg-resources==0.0.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
